requirements.txt
pytest==7.3.1
selenium==4.9.1
Appium-Python-Client==2.10.1